requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.2.7
beautifulsoup4==4.9.1
billiard==3.6.4.0
celery==5.0.5
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
comtypes==1.1.10
cryptography==3.4.7
cycler==0.10.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
Django==2.2
django-admin-honeypot==1.1.0
django-admin-interface==0.16.1
django-allauth==0.44.0
django-cleanup==5.2.0
django-colorfield==0.4.1
django-contactform==2.2.0
django-contactforms==2.0
django-cors-headers==3.7.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-debug-toolbar==3.2
django-defender==0.8.0
django-enumfields==2.1.1
django-extensions==3.1.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.5.0
django-maintenance-mode==0.16.0
django-model-utils==4.1.1
django-otp==1.0.3
django-phone-field==1.8.1
django-ranged-response==0.2.0
django-rest-registration==0.7.0
django-simple-captcha==0.5.14
django-social-share==2.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
et-xmlfile==1.0.1
filelock==3.0.12
futures==3.1.1
googletrans==3.0.0
goslate==1.5.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.15
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.9
jsonfield==3.1.0
kiwisolver==1.2.0
kombu==5.0.2
MarkupPy==1.14
matplotlib==3.3.0
MouseInfo==0.1.3
names==0.3.0
numpy==1.19.5
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.4.0
plyer==1.4.3
prompt-toolkit==3.0.18
psutil==5.7.2
psycopg2-binary==2.8.6
py-notifier==0.1.0
PyAutoGUI==0.9.52
pycparser==2.20
PyDictionary==1.5.2
PyGetWindow==0.0.9
PyJWT==2.0.1
pymongo==3.12.1
PyMsgBox==1.0.9
pyobjc==7.2
pyobjc-core==7.2
pyobjc-framework-Accounts==7.2
pyobjc-framework-AddressBook==7.2
pyobjc-framework-AdSupport==7.2
pyobjc-framework-AppleScriptKit==7.2
pyobjc-framework-AppleScriptObjC==7.2
pyobjc-framework-ApplicationServices==7.2
pyobjc-framework-AuthenticationServices==7.2
pyobjc-framework-AutomaticAssessmentConfiguration==7.2
pyobjc-framework-Automator==7.2
pyobjc-framework-AVFoundation==7.2
pyobjc-framework-AVKit==7.2
pyobjc-framework-BusinessChat==7.2
pyobjc-framework-CalendarStore==7.2
pyobjc-framework-CFNetwork==7.2
pyobjc-framework-CloudKit==7.2
pyobjc-framework-Cocoa==7.2
pyobjc-framework-Collaboration==7.2
pyobjc-framework-ColorSync==7.2
pyobjc-framework-Contacts==7.2
pyobjc-framework-ContactsUI==7.2
pyobjc-framework-CoreAudio==7.2
pyobjc-framework-CoreAudioKit==7.2
pyobjc-framework-CoreBluetooth==7.2
pyobjc-framework-CoreData==7.2
pyobjc-framework-CoreHaptics==7.2
pyobjc-framework-CoreLocation==7.2
pyobjc-framework-CoreMedia==7.2
pyobjc-framework-CoreMediaIO==7.2
pyobjc-framework-CoreMIDI==7.2
pyobjc-framework-CoreML==7.2
pyobjc-framework-CoreMotion==7.2
pyobjc-framework-CoreServices==7.2
pyobjc-framework-CoreSpotlight==7.2
pyobjc-framework-CoreText==7.2
pyobjc-framework-CoreWLAN==7.2
pyobjc-framework-CryptoTokenKit==7.2
pyobjc-framework-DeviceCheck==7.2
pyobjc-framework-DictionaryServices==7.2
pyobjc-framework-DiscRecording==7.2
pyobjc-framework-DiscRecordingUI==7.2
pyobjc-framework-DiskArbitration==7.2
pyobjc-framework-DVDPlayback==7.2
pyobjc-framework-EventKit==7.2
pyobjc-framework-ExceptionHandling==7.2
pyobjc-framework-ExecutionPolicy==7.2
pyobjc-framework-ExternalAccessory==7.2
pyobjc-framework-FileProvider==7.2
pyobjc-framework-FileProviderUI==7.2
pyobjc-framework-FinderSync==7.2
pyobjc-framework-FSEvents==7.2
pyobjc-framework-GameCenter==7.2
pyobjc-framework-GameController==7.2
pyobjc-framework-GameKit==7.2
pyobjc-framework-GameplayKit==7.2
pyobjc-framework-ImageCaptureCore==7.2
pyobjc-framework-IMServicePlugIn==7.2
pyobjc-framework-InputMethodKit==7.2
pyobjc-framework-InstallerPlugins==7.2
pyobjc-framework-InstantMessage==7.2
pyobjc-framework-Intents==7.2
pyobjc-framework-IOSurface==7.2
pyobjc-framework-iTunesLibrary==7.2
pyobjc-framework-LatentSemanticMapping==7.2
pyobjc-framework-LaunchServices==7.2
pyobjc-framework-libdispatch==7.2
pyobjc-framework-LinkPresentation==7.2
pyobjc-framework-LocalAuthentication==7.2
pyobjc-framework-MapKit==7.2
pyobjc-framework-MediaAccessibility==7.2
pyobjc-framework-MediaLibrary==7.2
pyobjc-framework-MediaPlayer==7.2
pyobjc-framework-MediaToolbox==7.2
pyobjc-framework-Metal==7.2
pyobjc-framework-MetalKit==7.2
pyobjc-framework-MetalPerformanceShaders==7.2
pyobjc-framework-ModelIO==7.2
pyobjc-framework-MultipeerConnectivity==7.2
pyobjc-framework-NaturalLanguage==7.2
pyobjc-framework-NetFS==7.2
pyobjc-framework-Network==7.2
pyobjc-framework-NetworkExtension==7.2
pyobjc-framework-NotificationCenter==7.2
pyobjc-framework-OpenDirectory==7.2
pyobjc-framework-OSAKit==7.2
pyobjc-framework-OSLog==7.2
pyobjc-framework-PencilKit==7.2
pyobjc-framework-Photos==7.2
pyobjc-framework-PhotosUI==7.2
pyobjc-framework-PreferencePanes==7.2
pyobjc-framework-PushKit==7.2
pyobjc-framework-Quartz==7.2
pyobjc-framework-QuickLookThumbnailing==7.2
pyobjc-framework-SafariServices==7.2
pyobjc-framework-SceneKit==7.2
pyobjc-framework-ScreenSaver==7.2
pyobjc-framework-ScriptingBridge==7.2
pyobjc-framework-SearchKit==7.2
pyobjc-framework-Security==7.2
pyobjc-framework-SecurityFoundation==7.2
pyobjc-framework-SecurityInterface==7.2
pyobjc-framework-ServiceManagement==7.2
pyobjc-framework-Social==7.2
pyobjc-framework-SoundAnalysis==7.2
pyobjc-framework-Speech==7.2
pyobjc-framework-SpriteKit==7.2
pyobjc-framework-StoreKit==7.2
pyobjc-framework-SyncServices==7.2
pyobjc-framework-SystemConfiguration==7.2
pyobjc-framework-SystemExtensions==7.2
pyobjc-framework-UserNotifications==7.2
pyobjc-framework-VideoSubscriberAccount==7.2
pyobjc-framework-VideoToolbox==7.2
pyobjc-framework-Vision==7.2
pyobjc-framework-WebKit==7.2
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
pyshorteners==1.0.1
python-barcode==0.13.1
python-dateutil==2.8.1
python3-openid==3.2.0
pyttsx3==2.90
pytube==9.6.0
pytube3==9.6.4
PyTweening==1.0.3
pytz==2021.3
PyYAML==5.4.1
qrcode==6.1
razorpay==1.2.0
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rubicon-objc==0.4.0
sentry-sdk==1.0.0
six==1.15.0
sniffio==1.1.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
tablib==3.0.0
touch==2020.12.3
typing-extensions==3.7.4.2
urllib3==1.25.9
values==2020.12.3
vine==5.0.0
virtualenv==20.4.6
waitress==1.4.4
wcwidth==0.2.5
whitenoise==5.2.0
wikipedia==1.4.0
xlrd==2.0.1
xlwt==1.3.0
